updated version (working on check && mate)

+++ SchaakStuk.cpp
@@ -12,27 +12,17 @@ vector<pair<int,int>> SchaakStuk::geldige_zetten(Game& g){
     return zetten;
 }
 
-bool SchaakStuk::check_danger() {
-    for(const auto& pair : danger_posities){
-        if(pair.first == rij && pair.second == kolom){
-            return true;
-        }
-    }
-    return false;
-}
-
 vector<pair<int,int>> Pion::geldige_zetten(Game& g){
     vector<pair<int,int>> zetten;
     int rij = this->rij;
     int kolom = this->kolom;
 
     if(g.schaakboord[rij][kolom]->getKleur() == zwart){ // check black
-
-        if(g.schaakboord[rij+1][kolom+1] != nullptr && g.schaakboord[rij+1][kolom+1]->getKleur() == wit && g.schaakboord[rij][kolom]->kolom < 7) {
+        if(g.schaakboord[rij+1][kolom+1] != nullptr && g.schaakboord[rij+1][kolom+1]->getKleur() == wit && g.schaakboord[rij][kolom]->kolom < 7 && rij+1 >= 0 && rij+1 <= 7) {
             zetten.push_back(make_pair(rij + 1, kolom + 1)); // kill schijn rechts indien niet op kolom 7
             danger_posities.push_back(make_pair(rij+1, kolom+1)); //DEZE POSITIE KAN GEKILLED WORDEN
         }
-        if(g.schaakboord[rij+1][kolom-1] != nullptr && g.schaakboord[rij+1][kolom-1]->getKleur() == wit && g.schaakboord[rij][kolom]->kolom > 0){
+        if(g.schaakboord[rij+1][kolom-1] != nullptr && g.schaakboord[rij+1][kolom-1]->getKleur() == wit && g.schaakboord[rij][kolom]->kolom > 0 && rij+1 >= 0 && rij +1 <= 7){
             zetten.push_back(make_pair(rij+1,kolom-1)); // kill schijn links indien niet op kolom 0
             danger_posities.push_back(make_pair(rij+1, kolom-1)); //DEZE POSITIE KAN GEKILLED WORDEN
         }
@@ -45,20 +35,16 @@ vector<pair<int,int>> Pion::geldige_zetten(Game& g){
             zetten.push_back(make_pair(rij+2, kolom));
         }
     } else if(g.schaakboord[rij][kolom]->getKleur() == wit){ // check white
-        if(g.schaakboord[rij-1][kolom+1] != nullptr && g.schaakboord[rij-1][kolom+1]->getKleur() == zwart && g.schaakboord[rij][kolom]->kolom < 7) {
+        if(g.schaakboord[rij-1][kolom+1] != nullptr && g.schaakboord[rij-1][kolom+1]->getKleur() == zwart && g.schaakboord[rij][kolom]->kolom < 7 && rij-1 >= 0 && rij-1 <= 7) {
             zetten.push_back(make_pair(rij-1, kolom + 1)); // kill schijn rechts idien kolom < 7
             danger_posities.push_back(make_pair(rij-1, kolom+1)); //DEZE POSITIE KAN GEKILLED WORDEN
         }
-        if(g.schaakboord[rij-1][kolom-1] != nullptr && g.schaakboord[rij-1][kolom-1]->getKleur() == zwart && g.schaakboord[rij][kolom]->kolom > 0) {
+        if(g.schaakboord[rij-1][kolom-1] != nullptr && g.schaakboord[rij-1][kolom-1]->getKleur() == zwart && g.schaakboord[rij][kolom]->kolom > 0 && rij-1 >= 0 && rij-1 <= 7) {
             zetten.push_back(make_pair(rij-1, kolom - 1)); // kill schijn links idien kolom > 0
             danger_posities.push_back(make_pair(rij-1, kolom-1)); //DEZE POSITIE KAN GEKILLED WORDEN
         }
         if(g.schaakboord[rij-1][kolom] == nullptr){
             zetten.push_back(make_pair(rij-1, kolom));
-        } else if(g.schaakboord[rij][kolom-1] != nullptr){
-            if(g.schaakboord[rij][kolom-1]->getKleur() == zwart){
-                zetten.push_back(make_pair(rij, kolom-1));
-            }
         } else{
             return zetten;
         }
@@ -612,7 +598,7 @@ vector<pair<int,int>> Koning::geldige_zetten(Game& g){
         //down mid
         if(rij+1 < 8){
             if(g.schaakboord[rij+1][kolom] == nullptr){
-                zetten.push_back(make_pair(rij+1, kolom));
+                zetten.push_back(make_pair(rij-1,kolom));
             } else if(g.schaakboord[rij+1][kolom] != nullptr && g.schaakboord[rij+1][kolom]->getKleur() == wit){
                 zetten.push_back(make_pair(rij+1, kolom));
                 danger_posities.push_back(make_pair(rij+1, kolom)); //DEZE POSITIE KAN GEKILLED WORDEN
